Tidy up.
- Remove lots of as-casts
- Restructure into folder
